Model improvements and feature importance
I did some improvements to the models. Chose only the Random Forest and the Naive Bayes models because those were the best.
Saw the feature importance with 2 different methods that gave me 2 different groups of variables.
I created 4 models, 2 (RF and NB) for each group of variables.
I still need to choose witch is the best model to use and witch is the best group of variables to use.